# src/backend/backend/Features/Product/Models/ProductCategory.cs
namespace backend.Features.Product.Models;

// TODO: Add Subcategories
public enum ProductCategory
{
    Accessories,
    Hardware,
    Software,
    Books,
    Movies,
    Music,
    Clothes,
    Games,
    Other
}

public enum SubProductCategory
{
    
}