# src/client/swagger.json
{
...
  "swagger": "2.0",
  "info": {
    "title": "Shopping APi",
    "description": "Web API for Shopping",
    "contact": {
      "name": "systemad",
      "url": "https://github.com/Systemad/",
      "email": ""
    },
    "version": "v1"
  },
  "host": "localhost:7099",
  "schemes": ["https"],
  "paths": {
...
    "/category/{category}": {
      "get": {
        "tags": ["Category"],
        "summary": "Get items of a specific category, default is 10",
        "operationId": "Category_GetItemsForCategory",
        "produces": ["text/plain", "application/json", "text/json"],
        "parameters": [
          {
            "type": "string",
            "name": "category",
            "in": "path",
            "required": true,
            "description": "Enter a category to fetch items of",
            "x-schema": {
              "$ref": "#/definitions/ProductCategory"
            },
            "x-nullable": false,
            "enum": [
              "Accessories",
              "Hardware",
              "Software",
              "Books",
              "Movies",
              "Music",
              "Games",
              "Other"
            ]
          },
          {
            "type": "integer",
            "name": "limit",
            "in": "query",
            "description": "Enter amount, leave empty if default amount should be retrieved",
            "format": "int32",
            "default": 10,
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "List of products",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProductDetail"
              }
            }
          }
        }
      }
    },
    "/category/random": {
      "get": {
        "tags": ["Category"],
        "summary": "Get total items of a random category, default is 10",
        "operationId": "Category_GetItemsOfRandomCategory",
        "produces": ["text/plain", "application/json", "text/json"],
        "parameters": [
          {
            "type": "integer",
            "name": "limit",
            "in": "query",
            "description": "Enter amount, leave empty if default amount should be retrieved",
            "format": "int32",
            "default": 10,
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "List of products",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProductDetail"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "ProductDetail": {
      "type": "object",
      "required": [
        "id",
        "name",
        "description",
        "category",
        "quantity",
        "price",
        "imageUrl"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "category": {
          "$ref": "#/definitions/ProductCategory"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        },
        "price": {
          "type": "number",
          "format": "decimal"
        },
        "imageUrl": {
          "type": "string"
        }
      }
    },
    "ProductCategory": {
      "type": "string",
      "description": "",
      "x-enumNames": [
        "Accessories",
        "Hardware",
        "Software",
        "Books",
        "Movies",
        "Music",
        "Games",
        "Other"
      ],
      "enum": [
        "Accessories",
        "Hardware",
        "Software",
        "Books",
        "Movies",
        "Music",
        "Games",
        "Other"
      ]
    }
  }
}